+++ app/assets/packs/board.js
@@ -1,0 +1,4 @@
+import Board from "javascript/boards/show/board"
+import mount from "javascript/lib/mount"
+
+mount(Board, "board");