+++ app/controllers/pages_controller.rb
@@ -1,7 +1,13 @@
 class PagesController < ApplicationController
+  before_action :authenticate_user!, only: [:app]
+
   def home
   end
 
   def pricing
+  end
+
+  def dashboard
+    @boards = current_user.boards
   end
 end