+++ app/helpers/application_helper.rb
@@ -5,8 +5,8 @@ module ApplicationHelper
   end
 
   def avatar_url(user)
-      default_url = "#{root_url}images/guest.png"
-      gravatar_id = Digest::MD5.hexdigest(user.email.downcase)
-      "http://gravatar.com/avatar/#{gravatar_id}.png?s=48&d=#{CGI.escape(default_url)}"
-    end
+    default_url = "#{root_url}images/guest.png"
+    gravatar_id = Digest::MD5.hexdigest(user.email.downcase)
+    "http://gravatar.com/avatar/#{gravatar_id}.png?s=48&d=#{CGI.escape(default_url)}"
+  end
 end