+++ app/models/list.rb
@@ -1,0 +1,4 @@
+class List < ApplicationRecord
+  belongs_to :board
+  has_many :cards, -> { order(position: :asc) }, dependent: :delete_all
+end