+++ db/migrate/20200517013057_create_lists.rb
@@ -1,0 +1,10 @@
+class CreateLists < ActiveRecord::Migration[6.0]
+  def change
+    create_table :lists do |t|
+      t.references :board, null: false, foreign_key: true
+      t.string :title
+
+      t.timestamps
+    end
+  end
+end