+++ db/migrate/20200604205336_change_order_to_position.rb
@@ -1,0 +1,5 @@
+class ChangeOrderToPosition < ActiveRecord::Migration[6.0]
+  def change
+    rename_column :cards, :order, :position
+  end
+end